runtime/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.40.0
httpx>=0.27.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.9.0
cryptography>=43.0.0
python-multipart>=0.0.12
